Satellite auth
