+++ src/components/Auth/clerk.tsx
@@ -11,11 +11,6 @@ export const Clerk = ({ children }: Props) => {
     <ClerkProvider
       isSatellite={process.env.NEXT_PUBLIC_CLERK_IS_SATELLITE === "true"}
       domain={(url) => url.host}
-      signInUrl={
-        process.env.NEXT_PUBLIC_SIGN_IN_URL
-          ? process.env.NEXT_PUBLIC_SIGN_IN_URL
-          : undefined
-      }
       publishableKey={process.env.NEXT_PUBLIC_CLERK_PUBLISHABLE_KEY}
     >
       {children}
